% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{agreaterb}
\alias{agreaterb}
\title{Check if event time is larger than other event time}
\usage{
agreaterb(a, b)
}
\arguments{
\item{a}{Vector of event times}

\item{b}{Vector of event times}
}
\value{
Matrix of size (length(a) * length(b)) with binary values indicating 
whether the event times in a are larger than the event times in b
}
\description{
Function which takes as input a vector a with event times 
and a vector b with event times and checks whether each entry in a is 
larger  than the entries in b.
}
