% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{ainB}
\alias{ainB}
\title{Check if event time is contained within half-open interval}
\usage{
ainB(a, B)
}
\arguments{
\item{a}{Vector of event times}

\item{B}{Two column matrix containing intervals}
}
\value{
Matrix of size (length(a) * nrow(B)) with binary values indicating 
whether the event times in a are contained in the intervals of B
}
\description{
Function which takes as input a vector a with event times 
and a 2 column matrix B representing half-open intervals (l, R] and checks whether 
each event time is contained in each half-open interval.
}
