% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{Aintersectb}
\alias{Aintersectb}
\title{Check if half-open intervals intersect with event times}
\usage{
Aintersectb(A, b, A.left.open = FALSE)
}
\arguments{
\item{A}{Two column matrix containing intervals}

\item{b}{Vector of event times}

\item{A.left.open}{Are the intervals in A open on the left side? Default = FALSE.}
}
\value{
Numeric vector of event times from b that intersect with A.
}
\description{
Function which takes as input a 2 column matrix of half-open 
intervals and a vector of event times and returns the event times that 
intersect with the half-open intervals.
}
