% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{Alargerb}
\alias{Alargerb}
\title{Check if closed interval is contained in half-open infinite interval}
\usage{
Alargerb(A, b)
}
\arguments{
\item{A}{Two column matrix containing intervals to be checked for being 
contained in b}

\item{b}{Vector indicating left point in corresponding [b, Infinity) interval}
}
\value{
Matrix of size (nrow(A) * length(b)) with binary values indicating 
whether the intervals in A are contained in the ones induced by b
}
\description{
Function which takes as input a matrix with 2 columns and a 
vector indicating left points of intervals [b, Infinity)
and checks whether each interval in the matrix is contained within 
the corresponding interval derived from b.
}
