% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary_search_larger}
\alias{binary_search_larger}
\title{Binary search - Larger}
\usage{
binary_search_larger(v, data)
}
\arguments{
\item{v}{- sorted vector instance}

\item{data}{- value to search}
}
\value{
index of first value larger than data -1, -1 otherwise
}
\description{
Find index of first value larger than x in a sorted array
}
\keyword{internal}
