% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary_search_larger_equal}
\alias{binary_search_larger_equal}
\title{Binary search - Larger or equal to}
\usage{
binary_search_larger_equal(v, data)
}
\arguments{
\item{v}{- sorted vector instance}

\item{data}{- value to search}
}
\value{
index of first value larger or equal than data, -1 otherwise
}
\description{
Find index of first value larger or equal to x in a sorted array
}
\keyword{internal}
