% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_multinomial.R
\name{EM_multinomial}
\alias{EM_multinomial}
\title{Helper function for npmsm()}
\usage{
EM_multinomial(
  gd,
  tmat,
  tmat2,
  support_manual,
  exact,
  maxit,
  tol,
  conv_crit,
  manual,
  verbose,
  newmet,
  include_inf,
  checkMLE,
  checkMLE_tol,
  prob_tol,
  remove_bins,
  init_int = init_int,
  ...
)
}
\arguments{
\item{gd}{A \code{data.frame} with the following named columns
\describe{
  \item{\code{id}:}{Subject idenitifier;}
  \item{\code{state}:}{State at which the subject is observed at \code{time};}
  \item{\code{time}:}{Time at which the subject is observed;}
} The true transition time between states is then interval censored between the times.}

\item{tmat}{A transition matrix as created by \code{transMat}}

\item{support_manual}{Used for specifying a manual support region for the transitions.
A list of length the number of transitions in \code{tmat}, 
each list element containing a data frame with 2 names columns L and R indicating the 
left and right values of the support intervals}

\item{exact}{Numeric vector indicating to which states transitions are observed at exact times.
Must coincide with the column number in \code{tmat}.}

\item{maxit}{Maximum number of iterations.}

\item{tol}{Tolerance of the procedure.}

\item{conv_crit}{Convergence criterion. Stops procedure when the difference 
in the chosen quantity between two consecutive iterations is smaller 
than the tolerance level \code{tol}. One of the following:
\describe{
\item{"haz"}{Stop when change in estimated intensities (hazards) \code{< tol}.}
\item{"prob"}{Stop when change in estimated probabilities \code{< tol}.}
\item{"lik"}{Stop when change in observed-data likelihood \code{< tol}.}
} Default is "haz". The options "haz" and "lik" can be compared across different
\code{method}s, but "prob" is dependent on the chosen \code{method}. Most 
conservative (safest) is "prob", followed by "haz" and finally "lik".}

\item{manual}{Manually specify starting transition intensities?}

\item{verbose}{Should iteration messages be printed? Default is FALSE}

\item{newmet}{Should contributions after last observation time also be used 
in the likelihood? Default is FALSE.}

\item{include_inf}{Should an additional bin from the largest observed time to 
infinity be included in the algorithm? Default is FALSE.}

\item{checkMLE}{Should a check be performed whether the estimate has converged 
towards a true Maximum Likelihood Estimate? Default is TRUE.}

\item{checkMLE_tol}{Tolerance for checking whether the estimate has converged to MLE.
Whenever an estimated transition intensity is smaller than the tolerance, it is assumed 
to be zero.}

\item{prob_tol}{If an estimated probability is smaller than \code{prob_tol}, 
it will be set to zero during estimation. Default value is \code{tol}.}

\item{remove_bins}{Should bins be removed during the algorithm if there is 
0 estimated intensity in all of the transitions? Significantly improves 
computation speed for large data sets. Note that 0 means the estimated intensities 
are smaller than \code{prob_tol}. Default is FALSE.}

\item{init_int}{A vector of length 2, with the first entry indicating what 
percentage of mass should be distributed over (second entry) what percentage 
of all first bins. Default is c(0, 0), in which case the argument is ignored.
This argument has no practical uses and only exists for demonstration purposes.}

\item{...}{Not used yet}
}
\description{
For a general Markov chain multi-state model with interval censored 
transitions calculate the NPMLE using an EM algorithm with multinomial approach
}
\references{
Michael G. Hudgens, On Nonparametric Maximum Likelihood Estimation with 
Interval Censoring and Left Truncation, Journal of the Royal Statistical Society 
Series B: Statistical Methodology, Volume 67, Issue 4, September 2005, Pages 573–587,
 \doi{10.1111/j.1467-9868.2005.00516.x}
}
\keyword{internal}
