% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep_smooth.R
\name{Estep_smooth}
\alias{Estep_smooth}
\title{E-step of the EM algorithm for smooth estimation of transition intensities}
\usage{
Estep_smooth(fix_pars, subject_slices, EM_est, it_num)
}
\description{
E-step of the EM algorithm for smooth estimation of transition intensities
}
\keyword{internal}
