% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_intensity_matrices_cpp}
\alias{get_intensity_matrices_cpp}
\title{Extract intensity matrices from msfit object}
\usage{
get_intensity_matrices_cpp(object)
}
\arguments{
\item{object}{List containing msfit object with elements: Haz (DataFrame) and trans (IntegerMatrix)}
}
\value{
List containing intensity_matrices (3D array) and unique_times
}
\description{
Extract intensity matrices from msfit object
}
