% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep_smooth.R
\name{Mstep_smooth}
\alias{Mstep_smooth}
\title{Function for fitting proportional hazard model with baseline hazard}
\usage{
Mstep_smooth(fix_pars, EM_est, transno, from, Pen = Pen)
}
\arguments{
\item{fix_pars}{A list of fixed parameters during the EM procedure}

\item{EM_est}{A list of estimated quantities that change during the EM procedure}

\item{transno}{Transition number for the maximization procedure}

\item{from}{The state from which the relevant transition is taking place.}

\item{Pen}{Penalization matrix for the B-splines and regression coefficients.}
}
\value{
An object with fields:
\code{H} = hazards (matrix),
\code{cbx} = coefficient estimates (vector),
\code{lambda} = proposal for new lambda,
\code{ed} = effective dimension,
\code{G} = G matrix,
\code{ll} = log-likelihood,
\code{pen} = penalized part of log-likelihood,
\code{Mpen} = penalized M matrix
}
\description{
Function for fitting proportional hazard model with baseline hazard
}
\keyword{internal}
