% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npmsm.R
\name{npmsm}
\alias{npmsm}
\title{NPMLE for general multi-state model with interval censored transitions}
\usage{
npmsm(
  gd,
  tmat,
  method = c("binomial", "poisson"),
  support_manual,
  exact,
  maxit = 100,
  tol = 1e-04,
  conv_crit = c("haz", "prob", "lik"),
  verbose = FALSE,
  manual = FALSE,
  newmet = FALSE,
  include_inf = FALSE,
  checkMLE = TRUE,
  checkMLE_tol = 1e-10,
  prob_tol = tol,
  remove_redundant = TRUE,
  remove_bins = FALSE,
  estimateSupport = FALSE,
  init_int = c(0, 0),
  ...
)
}
\arguments{
\item{gd}{A \code{data.frame} with the following named columns
\describe{
  \item{\code{id}:}{Subject idenitifier;}
  \item{\code{state}:}{State at which the subject is observed at \code{time};}
  \item{\code{time}:}{Time at which the subject is observed;}
} The true transition time between states is then interval censored between the times.}

\item{tmat}{A transition matrix as created by \code{transMat}}

\item{method}{Which method should be used for the EM algorithm. Choices are 
\code{c("binomial", "poisson")}, with binomial the default.}

\item{support_manual}{Used for specifying a manual support region for the transitions.
A list of length the number of transitions in \code{tmat}, 
each list element containing a data frame with 2 names columns L and R indicating the 
left and right values of the support intervals}

\item{exact}{Numeric vector indicating to which states transitions are observed at exact times.
Must coincide with the column number in \code{tmat}.}

\item{maxit}{Maximum number of iterations.}

\item{tol}{Tolerance of the procedure. An increase in an iteration in the 
value of \code{conv_crit} of less than \code{tol} will make the procedure stop.}

\item{conv_crit}{Convergence criterion. Stops procedure when the difference 
in the chosen quantity between two consecutive iterations is smaller 
than the tolerance level \code{tol}. One of the following:
\describe{
\item{"haz"}{Stop when change in estimated intensities (hazards) \code{< tol}.}
\item{"prob"}{Stop when change in estimated probabilities \code{< tol}.}
\item{"lik"}{Stop when change in observed-data likelihood \code{< tol}.}
} Default is "haz". The options "haz" and "lik" can be compared across different
\code{method}s, but "prob" is dependent on the chosen \code{method}. Most 
conservative (safest) is "prob", followed by "haz" and finally "lik".}

\item{verbose}{Should iteration messages be printed? Default is FALSE}

\item{manual}{Manually specify starting transition intensities?}

\item{newmet}{Should contributions after last observation time also be used 
in the likelihood? Default is FALSE.}

\item{include_inf}{Should an additional bin from the largest observed time to 
infinity be included in the algorithm? Default is FALSE.}

\item{checkMLE}{Should a check be performed whether the estimate has converged 
towards a true Maximum Likelihood Estimate? Default is TRUE.}

\item{checkMLE_tol}{Tolerance for checking whether the estimate has converged to MLE.
Whenever an estimated transition intensity is smaller than the tolerance, it is assumed 
to be zero.}

\item{prob_tol}{If an estimated probability is smaller than \code{prob_tol}, 
it will be set to zero during estimation. Default value is \code{tol}.}

\item{remove_redundant}{Should redundant observations be removed before running 
the algorithm? Default is TRUE.}

\item{remove_bins}{Should bins be removed during the algorithm if there is 
0 estimated intensity in all of the transitions? Significantly improves 
computation speed for large data sets. Note that 0 means the estimated intensities 
are smaller than \code{prob_tol}. Default is FALSE.}

\item{estimateSupport}{Should the support of the transitions be estimated using 
the result of Hudgens (2005)? Currently produces incorrect support sets - 
DO NOT USE.}

\item{init_int}{A vector of length 2, with the first entry indicating what 
percentage of mass should be distributed over (second entry) what percentage 
of all first bins. Default is c(0, 0), in which case the argument is ignored.
This argument has no practical uses and only exists for demonstration purposes.}

\item{...}{Further arguments to \code{\link{estimate_support_msm}}}
}
\description{
For a general Markov chain multi-state model with interval censored 
transitions calculate the NPMLE.
}
\examples{
if(require(mstate)){
  tmat <- mstate::trans.illdeath()
}

eval_times <- function(n_obs, stop_time){
  cumsum( c( 0,  runif( n_obs-1, 0, 2*(stop_time-4)/(n_obs-1) ) ) )
}

sim_dat <- sim_id_weib(n = 50, n_obs = 6, stop_time = 15, eval_times = eval_times,
start_state = "stable", shape = c(0.5, 0.5, 2), scale = c(5, 10, 10/gamma(1.5)))

tmat <- mstate::trans.illdeath()
mod_fit <- npmsm(gd = sim_dat, tmat = tmat, tol = 1e-2)
plot(mod_fit$A)

}
\references{
Michael G. Hudgens, On Nonparametric Maximum Likelihood Estimation with 
Interval Censoring and Left Truncation, Journal of the Royal Statistical Society 
Series B: Statistical Methodology, Volume 67, Issue 4, September 2005, Pages 573–587,
 \doi{10.1111/j.1467-9868.2005.00516.x}
}
