% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.npmsm}
\alias{plot.npmsm}
\title{Plot a "npmsm" object}
\usage{
\method{plot}{npmsm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"npmsm"}}

\item{...}{Additional arguments to \code{\link[mstate:msfit]{msfit}}}
}
\value{
A plot will be produced in the plotting window
}
\description{
A wrapper for \code{\link[mstate:msfit]{msfit}}
}
