% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.smoothmsm}
\alias{plot.smoothmsm}
\title{Plot a "smoothmsm" object}
\usage{
\method{plot}{smoothmsm}(x, haz = c("haz", "cumhaz", "loghaz"), ...)
}
\arguments{
\item{x}{An object of class \code{"smoothmsm"}}

\item{haz}{Type of hazard to plot. Choices are \code{"haz"} (default: hazard),
\code{"cumhaz"} (cumulative hazard), \code{"loghaz"} (log-hazard).}

\item{...}{Additional arguments to \code{\link[mstate:plot.msfit]{plot.msfit}}}
}
\value{
A plot will be produced in the plotting window
}
\description{
Plot the cumulative intensities of a \code{'smoothmsm'} objects.
For the cumulative hazard, a wrapper for \code{\link[mstate:plot.msfit]{plot.msfit}} from the
\code{\link[mstate:mstate]{mstate}} package.
}
