% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probtrans.R
\name{plot_probtrans}
\alias{plot_probtrans}
\title{Plot the transition probabilities for a fitted \code{npmsm} model}
\usage{
plot_probtrans(
  npmsmlist,
  from = NULL,
  to = NULL,
  transitions = NULL,
  landmark,
  interpolate = TRUE,
  facet = TRUE,
  times_interpol = NULL,
  c.legend = TRUE,
  c.names = NULL
)
}
\arguments{
\item{npmsmlist}{An "npmsm" object or a list containing multiple "npmsm" objects}

\item{from}{A numeric value indicating the state from which we consider the 
transition probabilities. Default is NULL, meaning we consider transition probabilities 
from all states from which a direct transition is possible.}

\item{to}{A numeric vector indicating to which states we consider the transition 
probabilities. Only states that can be reached from the \code{"from"} state are considered.}

\item{transitions}{A numeric vector indicating which transitions to consider (plot). 
Can only be used if \code{"from"} is not specified, as it only works for direct transitions.}

\item{landmark}{A landmark time indicating from which time on survival should be determined.
If missing, the smallest between the time in the first "npmsm" object or 0 will be used.}

\item{interpolate}{Should the cumulative hazard be linearly interpolated before 
determining transition probabilities? Default is TRUE.}

\item{facet}{Should the resulting plot be faceted (one panel per transition)? 
Default is TRUE.}

\item{times_interpol}{At which times should the cumulative hazard be interpolated?
Only necessary to specify if interpolate = TRUE.}

\item{c.legend}{Should legend be displayed for colour (different entries in 
\code{npmsmlist})? Default is TRUE.}

\item{c.names}{A character vector indicating the names to display in the legend.
These names should represent the entries in \code{npmsmlist.}. Default = \code{NULL}.}
}
\description{
For a fitted \code{npmsm} model plot the transition probabilities
from a certain state for all possible (direct and indirect) transitions.
}
\examples{
#Empty for now


}
