% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.npmsm}
\alias{print.npmsm}
\title{Print a "npmsm" object}
\usage{
\method{print}{npmsm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"npmsm"}}

\item{...}{Additional arguments to print}
}
\value{
A summary of the fitted model will be displayed in the console
}
\description{
Print some details of a \code{\link{npmsm}} fit
}
