% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_subj_methods.R
\name{print.summary.probtrans.subjects}
\alias{print.summary.probtrans.subjects}
\title{Print method for a summary.probtrans.subjects object}
\usage{
\method{print}{summary.probtrans.subjects}(x, complete = FALSE, ...)
}
\arguments{
\item{x}{Object of class 'summary.probtrans.subjects', to be printed}

\item{complete}{Whether or not the complete estimated transition
probabilities should be printed (\code{TRUE}) or not (\code{FALSE}); default
is \code{FALSE}, in which case the estimated transition probilities will be
printed for the first and last 6 time points of each starting state or of
the selected times (or all when there are at most 12 of these time points}

\item{\dots}{Further arguments to print}
}
\description{
Print method for a summary.probtrans.subjects object
}
\examples{

\dontrun{
# If all time points should be printed, specify complete=TRUE in the print statement
print(x, complete=TRUE)
}

}
