% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_coxph.R
\name{probtrans_coxph2}
\alias{probtrans_coxph2}
\title{Second (hopefully faster) version of probtrans coxph}
\usage{
probtrans_coxph2(
  object,
  predt,
  direction = c("forward", "fixedhorizon"),
  newdata,
  trans
)
}
\description{
Second (hopefully faster) version of probtrans coxph
}
\keyword{internal}
