% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probtrans_D_cpp}
\alias{probtrans_D_cpp}
\title{Calculate transition probabilities (main function)}
\usage{
probtrans_D_cpp(int_mat, predt, direction = "forward", as_df = FALSE)
}
\arguments{
\item{int_mat}{List from get_intensity_matrices_cpp}

\item{predt}{Prediction time}

\item{direction}{String: "forward" or "fixedhorizon"}

\item{as_df}{Boolean: return as dataframe format}
}
\value{
Array or List of transition probabilities
}
\description{
Calculate transition probabilities (main function)
}
