% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_msm.R
\name{probtrans_msm}
\alias{probtrans_msm}
\title{Given a \code{msm} object, determine the transition probabilities (forward) 
at the specified times, starting from the smallest specified time.}
\usage{
probtrans_msm(msm, times)
}
\arguments{
\item{msm}{A \code{msm} object.}

\item{times}{A vector of times at which the transition probabilities should 
be determined.}
}
\value{
A \code{probtrans} object containing the estimated transition probabilities.
}
\description{
The main use of this function is to extract transition 
probabilities (as in \code{\link[mstate:probtrans]{probtrans}})  from an 
\code{\link[msm:msm]{msm}} object. Functions from the \code{mstate} package
can then be used for further inspection.
}
