% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{prod_lambda_G_base}
\alias{prod_lambda_G_base}
\title{Calculate the product of intensities over interval decided by 
failure times}
\usage{
prod_lambda_G_base(lambda, t_n, Q, A)
}
\arguments{
\item{lambda}{Intensities of the 2->3 transition}

\item{t_n}{Unique failure times, same length as lambda}

\item{Q}{Matrix (2 column) containing support intervals as rows}

\item{A}{Matrix (2 column) containing censoring intervals as rows}
}
\description{
This function calculates \eqn{\prod_{\lambda}}{prod(\lambda)} G as defined in 
Frydman (1995), with \eqn{t_n}{t_n} the failure times. Note that length(t_n) must be 
equal to length(lambda)
}
