% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redundant_observations.R
\name{remove_redundant_observations}
\alias{remove_redundant_observations}
\title{Remove redundant observations from supplied data frame}
\usage{
remove_redundant_observations(gd, tmat)
}
\arguments{
\item{gd}{A \code{data.frame} with the following named columns
\describe{
  \item{\code{id}:}{Subject idenitifier;}
  \item{\code{state}:}{State at which the subject is observed at \code{time};}
  \item{\code{time}:}{Time at which the subject is observed;}
} The true transition time between states is then interval censored between the times.}

\item{tmat}{A transition matrix as created by \code{transMat}}
}
\description{
Remove redundant observed states from a supplied data frame. 
Observations are redundant either when we observe an absorbing state multiple 
times (as we cannot leave an absorbing state), or 
when a transient state is observed multiple times between transitions (as 
we cannot have loops, therefore no extra information is provided when 
we observe a transient state multiple times).
}
