% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothmsm.R
\name{smoothmsm}
\alias{smoothmsm}
\title{Smooth hazard estimation for general multi-state model with interval censored
transitions}
\usage{
smoothmsm(
  gd,
  tmat,
  exact,
  formula,
  data,
  deg_splines = 3,
  ord_penalty = 2,
  n_intervals,
  maxit = 100,
  tol = 1e-04,
  conv_crit = c("haz", "prob", "lik"),
  verbose = FALSE,
  prob_tol = tol/10
)
}
\arguments{
\item{gd}{A \code{data.frame} with the following named columns
\describe{
  \item{\code{id}:}{Subject idenitifier;}
  \item{\code{state}:}{State at which the subject is observed at \code{time};}
  \item{\code{time}:}{Time at which the subject is observed;}
} The true transition time between states is then interval censored between the times.}

\item{tmat}{A transition matrix as created by \code{\link[mstate:transMat]{transMat}}}

\item{exact}{Numeric vector indicating to which states transitions are observed at exact times.
Must coincide with the column number in \code{tmat}.}

\item{formula}{Formula to interpret in data for covariates.}

\item{data}{A \code{data.frame} containing a column called \code{'id'} 
(identifying the subjects in \code{gd}) and 
variables which to interpret in \code{formula}.}

\item{deg_splines}{Degree to use for the B-spline basis functions. Defaults 
to cubic B-splines (degree 3).}

\item{ord_penalty}{Order of the P-spline penalty (penalty on the difference 
between d-order differences of spline coefficients). See Eilers \& Marx 
Section 2.3. Defaults to 2}

\item{maxit}{Maximum number of iterations. Default = 100.}

\item{tol}{Tolerance of the convergence procedure. A change in the value of 
\code{conv_crit} in an iteration of less than \code{tol} will make the procedure stop.}

\item{conv_crit}{Convergence criterion. Stops procedure when the difference 
in the chosen quantity between two consecutive iterations is smaller 
than the tolerance level \code{tol}. One of the following:
\describe{
\item{"haz"}{Stop when change in maximum estimated intensities (hazards) \code{ < tol}.}
\item{"prob"}{Stop when change in estimated probabilities \code{ < tol}.}
\item{"lik"}{Stop when change in observed-data likelihood \code{ < tol}.}
} Default is "haz". The options "haz" and "lik" can be compared across different
\code{method}s, but "prob" is dependent on the chosen \code{method}. Most 
conservative (requiring most iterations) is "prob", followed by "haz" and finally "lik".}

\item{verbose}{Should iteration messages be printed? Default is FALSE}

\item{prob_tol}{If an estimated probability is smaller than \code{prob_tol}, 
it will be set to zero during estimation. Default value is \code{tol/10}.}

\item{inits}{Which distribution should be used to generate the initial estimates 
of the intensities in the EM algorithm. One of c("equalprob", "unif", "beta"), 
with "equalprob" assigning 1/K to each intensity, with K the number of distinct 
observation times (\code{length(unique(gd[, "time"]))}). For "unif", each 
intensity is sampled from the Unif[0,1] 
distribution and for "beta" each intensity is sampled from the Beta(a, b) distribution. 
If "beta" is chosen, the argument \code{beta_params} must be specified as a 
vector of length 2 containing the parameters of the beta distribution.
Default = "equalprob".}
}
\description{
For a general Markov chain multi-state model with interval censored 
transitions calculate the NPMLE of the transition intensities. The estimates 
are returned as an \code{\link[mstate:msfit]{msfit}} object.
}
\references{
Eilers, P.H.C. and Marx, B.D., Practical Smoothing: The Joys of P-splines, 
Cambridge University Press (2021)
}
