% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_coxph.R
\name{subject_specific_intensity_matrix}
\alias{subject_specific_intensity_matrix}
\title{Determine subject specific intensity matrix for a single subject}
\usage{
subject_specific_intensity_matrix(
  subject_specific_risk,
  intensity_matrices_zero_diag,
  trans,
  n_transitions,
  n_times
)
}
\description{
Determine subject specific intensity matrix for a single subject
}
\keyword{internal}
