% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_frydman.R
\name{support_frydman}
\alias{support_frydman}
\title{Find support for Illness-death model. Only 
applicable to Frydman (1995) setting!!}
\usage{
support_frydman(data_idx)
}
\arguments{
\item{data_idx}{List containing all the necessary data, output from the 
\code{\link{msm_frydman}} function.}
}
\value{
A: Matrix containing the intervals indicating when the transitions 
were censored.
unique_L_times: Union of failure times in Group 4 and censoring times in 
Group 1.
L_tilde: Left intervals for constructing the Q intervals
R_tilde: Right intervals for constructing the Q intervals
Q_mat: Matrix containing the support set for the 1->2 transition
I: number of intervals in the support set Q.
}
\description{
Find support for Illness-death model. Only 
applicable to Frydman (1995) setting!!
}
\details{
Finds the support sets Q for the interval censored transition
1 -> 2. The transitions 2->3 and 1->3 are exactly observed in this setting, 
therefore do not require determination of the support.
}
\references{
Frydman, H. (1995). Nonparametric Estimation of a Markov 
'Illness-Death' Process from Interval- Censored Observations, with 
Application to Diabetes Survival Data. Biometrika, 82(4), 773-789. 
\doi{10.2307/2337344}
}
\keyword{internal}
