% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_coxph.R
\name{trans_specific_risks}
\alias{trans_specific_risks}
\title{Calculate subject specific risks for subjects in newdata}
\usage{
trans_specific_risks(object, newdata, trans)
}
\description{
Return a 2 dimensional array, with subjects in the first dimension and
transition (numbers) in the second. Can expand later to include time-dependent
covariates by introducing extra dimension. Entries of the matrix are exp(lp)_i^m,
with i denoting the subject and m the transition.
}
\keyword{internal}
