% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_msm.R
\name{visualise_msm}
\alias{visualise_msm}
\title{Visualise data}
\usage{
visualise_msm(gd, npmsm, tmat, neat = TRUE, cutoff)
}
\arguments{
\item{gd}{A \code{data.frame} containing the following named columns:
\describe{
  \item{\code{id}:}{Identifier of subject;}
  \item{\code{state}:}{state of subject at \code{time};}
  \item{\code{time}:}{time at which subject is observed;}
}}

\item{npmsm}{Output from \code{\link{npmsm}} function}

\item{tmat}{A transition matrix as created by \code{transMat}}

\item{neat}{Boolean indicating whether redundant observations should be 
removed in the plot. Default is TRUE}

\item{cutoff}{cutoff value for numerically determining the support using
\code{\link{support_npmsm}}}
}
\description{
Visualise data
}
