% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_coxph.R
\name{baseline_intensities_from_coxmod}
\alias{baseline_intensities_from_coxmod}
\title{Recover baseline intensities (in the form of intensity_matrices) from a 
coxph() fit on multi-state data.}
\usage{
baseline_intensities_from_coxmod(object, tmat)
}
\description{
Recover baseline intensities (in the form of intensity_matrices) from a 
coxph() fit on multi-state data.
}
\keyword{internal}
