% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_coxph.R
\name{expand_covariates_long_data}
\alias{expand_covariates_long_data}
\title{Expand covariates for a data frame so that covariates can be transition 
specific.}
\usage{
expand_covariates_long_data(newdata)
}
\description{
Expand covariates for a data frame so that covariates can be transition 
specific.
}
\keyword{internal}
