% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep_smooth.R
\name{Mstep_smooth}
\alias{Mstep_smooth}
\title{Function for fitting proportional hazard model with baseline hazard}
\usage{
Mstep_smooth(fix_pars, EM_est, transno, from, Pen = Pen)
}
\arguments{
\item{Pen}{Penalty matrix}

\item{Y}{Expected events (matrix)}

\item{R}{Expected risk sets (matrix)}

\item{X}{Covariates (matrix)}

\item{B}{B-spline basis}

\item{lambda}{Smoothing parameter (number)}

\item{cbx}{Current coefficient estimates}
}
\value{
An object with fields:
\code{H} = hazards (matrix),
\code{cbx} = coefficient estimates (vector),
\code{lambda} = proposal for new lambda,
\code{ed} = effective dimension,
\code{G} = G matrix,
\code{ll} = log-likelihood,
\code{pen} = penalized part of log-likelihood,
\code{Mpen} = penalized M matrix
}
\description{
Function for fitting proportional hazard model with baseline hazard
}
