% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.smoothmsm}
\alias{plot.smoothmsm}
\title{Plot a "smoothmsm" object}
\usage{
\method{plot}{smoothmsm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"smoothmsm"}}

\item{...}{Additional arguments to \code{\link[base:plot]{plot}}}
}
\value{
A plot will be produced in the plotting window
}
\description{
Plot the cumulative intensities of a \code{'npmsm'} objects.
A wrapper for \code{\link[mstate:plot.msfit]{plot.msfit}} from the
\code{\link[mstate:mstate]{mstate}} package.
}
