% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv.R
\name{plot_surv}
\alias{plot_surv}
\title{Plot the transition specific survival probabilities for a fitted \code{npmsm} model}
\usage{
plot_surv(npmsmlist, landmark, support = FALSE, sup_cutoff = 1e-08)
}
\arguments{
\item{npmsmlist}{An "npmsm" object or a list containing multiple "npmsm" objects}

\item{landmark}{A landmark time indicating from which time on survival should be determined.
If missing, the smallest time in the first "npmsm" object will be used.}

\item{support}{Should the support regions be displayed as rectangles?}

\item{sup_cutoff}{Cutoff to be used for determining the support intervals.}
}
\description{
For a fitted \code{npmsm} model plot the transition specific 
survival probabilities. These are given by the product integral of the hazard 
increments estimated for a single transition. This is equivalent to a Kaplan-Meier 
estimator ignoring the existence of all other transitions.
}
