% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_coxph.R
\name{probtrans_coxph}
\alias{probtrans_coxph}
\title{Calculate subject specific transition probabilities from a 
multi-state \code{\link[survival:coxph]{coxph}} model.}
\usage{
probtrans_coxph(
  object,
  predt,
  direction = c("forward", "fixedhorizon"),
  newdata,
  trans
)
}
\arguments{
\item{object}{A \code{\link[survival:coxph]{coxph}} object fit on multi-state
data. Must contain a \code{strata(X)} term. Data used for the coxph() fit preferably prepared 
using \code{\link[mstate:msprep]{msprep}}.}

\item{predt}{A positive number indicating the prediction time. This is either 
the time at which the prediction is made (if \code{direction = "forward"}) or 
the time for which the prediction is to be made (if \code{direction = "backward"}).}

\item{direction}{One of \code{"forward"} (default) or \code{"fixedhorizon"},
indicating whether prediction is forward or for a fixed horizon}

\item{newdata}{A \code{data.frame} containing a single row for each transition per
subject in the data. For a model with m possible transitions, and n subjects 
\code{newdata} must have m*n rows. It must contain the following named columns:
\describe{
  \item{\code{id}:}{Unique identifier of the subject, can be numeric or character;}
  \item{\code{from}:}{State from which the transition takes place;}
  \item{\code{to}:}{State to which the transition takes place;}
  \item{\code{trans}:}{Transition number in the \code{'transMat'} \code{trans} this transition relates to;}
  \item{\code{"variables"}:}{The variables and their values for the subject
  identified by "id" for the transition this entry relates to. Names must 
  match the names of the variables in coxph \code{object}.}
} Note that newdata must contain a column containing the variable which was 
used to determine the stratum of a transition in \code{object}. 
Usually the stratum is determined from one of the required columns. The 
"variables" columns can usually be obtained using \code{\link[mstate:expand.covs]{expand.covs}}.}

\item{trans}{A transition matrix as created by \code{\link[mstate:transMat]{transMat}}.}
}
\value{
A list of length n (number of subjects in newdata), with each list component
an object of class \code{\link[mstate:probtrans]{probtrans}} for the associated 
subject.
}
\description{
Given a coxph model fit on multi-state data (prepared with \code{\link[mstate:msprep]{msprep}}),
determine transition probabilities for subjects in \code{newdata}.
}
\examples{
#Example from the mstate vignette
#We determine the subject specific transition probabilities for subjects
#in the ebmt3 data-set using only mstate functionality and using probtrans_coxph
if(require("mstate")){
  data(ebmt3)
  n <- nrow(ebmt3)
  tmat <- transMat(x = list(c(2, 3), c(3), c()), names = c("Tx",
                                                           "PR", "RelDeath"))
  ebmt3$prtime <- ebmt3$prtime/365.25
  ebmt3$rfstime <- ebmt3$rfstime/365.25
  covs <- c("dissub", "age", "drmatch", "tcd", "prtime")
  msbmt <- msprep(time = c(NA, "prtime", "rfstime"), status = c(NA,
                  "prstat", "rfsstat"), data = ebmt3, trans = tmat, keep = covs)
  #Expand covariates so that we can have transition specific covariates
  msbmt <- expand.covs(msbmt, covs, append = TRUE, longnames = FALSE)
  
  #Create extra variable to allow gender mismatch to have the same effect 
  #for transitions 2 and 3.
  msbmt$drmatch.2.3 <- msbmt$drmatch.2 + msbmt$drmatch.3
  
  #Introduce pr covariate for proportionality assumption of transitions 2 and 3
  #(only used in models 2 and 4)
  msbmt$pr <- 0
  msbmt$pr[msbmt$trans == 3] <- 1
  
  
  
  #-------------Models---------------------#
  
  #Simple model, transition specific covariates, each transition own baseline hazard
  c1 <- coxph(Surv(Tstart, Tstop, status) ~ dissub1.1 + dissub2.1 +
                age1.1 + age2.1 + drmatch.1 + tcd.1 + dissub1.2 + dissub2.2 +
                age1.2 + age2.2 + drmatch.2 + tcd.2 + dissub1.3 + dissub2.3 +
                age1.3 + age2.3 + drmatch.3 + tcd.3 + strata(trans), data = msbmt,
                method = "breslow")
  
  #Model with same baseline hazards for transitions 2 (1->3) and 3(2->3)
  #pr then gives the ratio of the 2 hazards for these transitions
  c2 <- coxph(Surv(Tstart, Tstop, status) ~ dissub1.1 + dissub2.1 +
                age1.1 + age2.1 + drmatch.1 + tcd.1 + dissub1.2 + dissub2.2 +
                age1.2 + age2.2 + drmatch.2 + tcd.2 + dissub1.3 + dissub2.3 +
                age1.3 + age2.3 + drmatch.3 + tcd.3 + pr + strata(to), data = msbmt,
                method = "breslow")
  
  #Same as c2, but now Gender mismatch has the same effect for both 
  c4 <- coxph(Surv(Tstart, Tstop, status) ~ dissub1.1 + dissub2.1 +
                age1.1 + age2.1 + drmatch.1 + tcd.1 + dissub1.2 + dissub2.2 +
                age1.2 + age2.2 + drmatch.2.3 + tcd.2 + dissub1.3 + dissub2.3 +
                age1.3 + age2.3  + tcd.3 + pr + strata(to), data = msbmt,
              method = "breslow")
              
  
  #We need to make a data.frame containing all subjects of interest
  ttmat <- to.trans2(tmat)[, c(2, 3, 1)]
  names(ttmat)[3] <- "trans"
  nd_n <- NULL
  for (j in 1:n) {
    # Select global covariates of subject j
    cllj <- ebmt3[j, covs]
    nd2 <- cbind(ttmat, rep(j, 3), rbind(cllj, cllj, cllj))
    colnames(nd2)[4] <- "id"
    # Make nd of class msdata to use expand.covs
    attr(nd2, "trans") <- tmat
    class(nd2) <- c("msdata", "data.frame")
    nd2 <- expand.covs(nd2, covs=covs, longnames = FALSE)
    nd2$drmatch.2.3 <- nd$drmatch.2 + nd$drmatch.3
    nd2$pr <- 0
    nd2$pr[nd$trans==3] <- 1
    nd2$strata <- c(1, 2, 2)
    nd_n <- rbind(nd_n, nd2)
   }
   
   icmstate_pt <- probtrans_coxph(c2, predt = 0, direction = "forward", 
                                  newdata = nd_n, trans = tmat)
                                  
   #Now we can plot the transition probabilities for each subject separately:
   plot(icmstate_pt[[1]])
   #icmstate_pt has length number of subjects in newdata
   #And icmstate_pt[[i]] is an object of class "probtrans", so you can 
   #use all probtrans functions: summary, plot etc.
}


}
