% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_npmsm.R
\name{support_npmsm}
\alias{support_npmsm}
\title{Numerically find the support of the transitions from a converged npmsm algorithm}
\usage{
support_npmsm(npmsm, cutoff = 1e-08)
}
\arguments{
\item{npmsm}{Output from \code{\link{npmsm}} function or an \code{\link[mstate:msfit]{msfit}} object.}

\item{cutoff}{Above which value is a mass in a bin considered to be non-zero? 
Default = 1e-8. Note that this is independent of bin size, so can be tricky!!}
}
\description{
For each transition in \code{tmat}, determine all consecutive 
bins with non-zero (higher than \code{cutoff}) transition intensities.
These then determine the numerical support of the transition.
}
